+++ app/Filament/Tables/SourceGroupsTable.php
@@ -2,6 +2,7 @@
 
 namespace App\Filament\Tables;
 
+use App\Filament\Tables\Traits\FiltersBySelection;
 use App\Models\Group;
 use App\Models\SourceGroup;
 use Filament\Actions\BulkActionGroup;
@@ -13,6 +14,8 @@ use Illuminate\Database\Query\Builder as QueryBuilder;
 
 class SourceGroupsTable
 {
+    use FiltersBySelection;
+
     public static function configure(Table $table): Table
     {
         return $table
@@ -103,61 +106,5 @@ class SourceGroupsTable
                     //
                 ]),
             ]);
-    }
-
-    private static function whereSelected(Builder $query, mixed $selectedValues, bool $selected): Builder
-    {
-        [$selectedIds, $selectedNames] = self::selectedIdsAndNames($selectedValues);
-
-        if (empty($selectedIds) && empty($selectedNames)) {
-            return $selected ? $query->whereRaw('1 = 0') : $query;
-        }
-
-        if (! $selected) {
-            return $query
-                ->when($selectedIds, fn (Builder $query): Builder => $query->whereNotIn('source_groups.id', $selectedIds))
-                ->when($selectedNames, fn (Builder $query): Builder => $query->whereNotIn('source_groups.name', $selectedNames));
-        }
-
-        return $query->where(function (Builder $query) use ($selectedIds, $selectedNames): void {
-            if (! empty($selectedIds)) {
-                $query->whereIn('source_groups.id', $selectedIds);
-            }
-
-            if (! empty($selectedNames)) {
-                $method = empty($selectedIds) ? 'whereIn' : 'orWhereIn';
-                $query->{$method}('source_groups.name', $selectedNames);
-            }
-        });
-    }
-
-    /**
-     * @return array{0: list<int>, 1: list<string>}
-     */
-    private static function selectedIdsAndNames(mixed $selectedValues): array
-    {
-        if (! is_array($selectedValues)) {
-            return [[], []];
-        }
-
-        $selectedIds = [];
-        $selectedNames = [];
-
-        foreach ($selectedValues as $value) {
-            if (is_numeric($value)) {
-                $selectedIds[] = (int) $value;
-
-                continue;
-            }
-
-            if (is_string($value) && $value !== '') {
-                $selectedNames[] = $value;
-            }
-        }
-
-        return [
-            array_values(array_unique($selectedIds)),
-            array_values(array_unique($selectedNames)),
-        ];
     }
 }
